requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
antropy=0.1.9=pypi_0
asttokens=3.0.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=24.3.0=pypi_0
ca-certificates=2024.12.31=h06a4308_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=pypi_0
et-xmlfile=2.0.0=pypi_0
eval-type-backport=0.2.2=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
filelock=3.17.0=pypi_0
fonttools=4.55.4=pypi_0
frozenlist=1.5.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
grpcio=1.69.0=pypi_0
huggingface-hub=0.27.1=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pypi_0
importlib-resources=6.5.2=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.5=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.5=pypi_0
joblib=1.4.2=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lightgbm=4.6.0=pypi_0
lightning=2.5.0.post0=pypi_0
lightning-utilities=0.11.9=pypi_0
llvmlite=0.43.0=pypi_0
lspopt=1.4.0=pypi_0
markdown=3.7=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mne=1.7.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.1.0=pypi_0
pip=24.2=py39h06a4308_0
platformdirs=4.3.6=pypi_0
pooch=1.8.2=pypi_0
prompt-toolkit=3.0.50=pypi_0
propcache=0.2.1=pypi_0
protobuf=5.29.3=pypi_0
psutil=6.1.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.1=pypi_0
python=3.9.21=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-graphviz=0.20.3=pypi_0
pytorch-lightning=2.5.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
safetensors=0.5.2=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
sentry-sdk=2.21.0=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=75.1.0=py39h06a4308_0
six=1.17.0=pypi_0
sleepecg=0.5.8=pypi_0
smmap=5.0.2=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.13.1=pypi_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorpac=0.6.5=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=0.9.16=pypi_0
tk=8.6.14=h39e8969_0
torch=2.2.2+cu121=pypi_0
torchmetrics=1.6.1=pypi_0
torchvision=0.17.2+cu121=pypi_0
torchviz=0.0.3=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
wandb=0.19.6=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py39h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.18.3=pypi_0
yasa=0.6.5=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
